requirements.txt
nonebot==1.9.1
psycopg2_binary==2.9.9
python-dotenv==1.0.1
Requests==2.32.3
